fix test

+++ core/src/main/java/org/elasticsearch/search/aggregations/MultiBucketConsumerService.java
@@ -71,7 +71,7 @@ public class MultiBucketConsumerService {
         @Override
         public void writeTo(StreamOutput out) throws IOException {
             super.writeTo(out);
-            out.writeLong(maxBuckets);
+            out.writeInt(maxBuckets);
         }
 
         public int getMaxBuckets() {
